+++ packages/rs-drive-abci/src/execution/validation/state_transition/state_transitions/identity_top_up/state/v0/mod.rs
@@ -6,6 +6,7 @@ use dpp::consensus::signature::{BasicECDSAError, SignatureError};
 use dpp::dashcore::signer;
 use dpp::dashcore::signer::double_sha;
 use dpp::identity::state_transition::AssetLockProved;
+use dpp::identity::KeyType;
 
 use dpp::prelude::ConsensusValidationResult;
 use dpp::serialization::Signable;
@@ -16,9 +17,13 @@ use dpp::version::PlatformVersion;
 use drive::state_transition_action::identity::identity_topup::IdentityTopUpTransitionAction;
 use drive::state_transition_action::StateTransitionAction;
 
-use drive::grovedb::TransactionArg;
 use crate::error::execution::ExecutionError;
-use crate::execution::validation::state_transition::common::asset_lock::proof::AssetLockProofStateValidation;
+use drive::grovedb::TransactionArg;
+
+use crate::execution::types::execution_operation::ExecutionOperation;
+use crate::execution::types::execution_operation::signature_verification_operation::SignatureVerificationOperation;
+use crate::execution::types::state_transition_execution_context::{StateTransitionExecutionContext, StateTransitionExecutionContextMethodsV0};
+use crate::execution::validation::state_transition::common::asset_lock::proof::validate::AssetLockProofValidation;
 use crate::execution::validation::state_transition::common::asset_lock::transaction::fetch_asset_lock_transaction_output_sync::fetch_asset_lock_transaction_output_sync;
 
 pub(in crate::execution::validation::state_transition::state_transitions::identity_top_up) trait IdentityTopUpStateTransitionStateValidationV0
@@ -26,6 +31,7 @@ pub(in crate::execution::validation::state_transition::state_transitions::identi
     fn validate_state_v0<C: CoreRPCLike>(
         &self,
         platform: &PlatformRef<C>,
+        execution_context: &mut StateTransitionExecutionContext,
         tx: TransactionArg,
         platform_version: &PlatformVersion,
     ) -> Result<ConsensusValidationResult<StateTransitionAction>, Error>;
@@ -33,6 +39,7 @@ pub(in crate::execution::validation::state_transition::state_transitions::identi
     fn transform_into_action_v0<C: CoreRPCLike>(
         &self,
         platform: &PlatformRef<C>,
+        execution_context: &mut StateTransitionExecutionContext,
         platform_version: &PlatformVersion,
     ) -> Result<ConsensusValidationResult<StateTransitionAction>, Error>;
 }
@@ -41,12 +48,13 @@ impl IdentityTopUpStateTransitionStateValidationV0 for IdentityTopUpTransition {
     fn validate_state_v0<C: CoreRPCLike>(
         &self,
         platform: &PlatformRef<C>,
+        execution_context: &mut StateTransitionExecutionContext,
         tx: TransactionArg,
         platform_version: &PlatformVersion,
     ) -> Result<ConsensusValidationResult<StateTransitionAction>, Error> {
         let mut validation_result = ConsensusValidationResult::<StateTransitionAction>::default();
 
-        validation_result.merge(self.asset_lock_proof().validate_state(
+        validation_result.merge(self.asset_lock_proof().validate(
             platform,
             tx,
             platform_version,
@@ -56,12 +64,13 @@ impl IdentityTopUpStateTransitionStateValidationV0 for IdentityTopUpTransition {
             return Ok(validation_result);
         }
 
-        self.transform_into_action_v0(platform, platform_version)
+        self.transform_into_action_v0(platform, execution_context, platform_version)
     }
 
     fn transform_into_action_v0<C: CoreRPCLike>(
         &self,
         platform: &PlatformRef<C>,
+        execution_context: &mut StateTransitionExecutionContext,
         platform_version: &PlatformVersion,
     ) -> Result<ConsensusValidationResult<StateTransitionAction>, Error> {
         let mut validation_result = ConsensusValidationResult::<StateTransitionAction>::default();
@@ -82,7 +91,7 @@ impl IdentityTopUpStateTransitionStateValidationV0 for IdentityTopUpTransition {
 
         // Verify one time signature
 
-        let singable_bytes = StateTransition::IdentityTopUp(self.clone()).signable_bytes()?;
+        let signable_bytes = StateTransition::IdentityTopUp(self.clone()).signable_bytes()?;
 
         let public_key_hash = tx_out
             .script_pubkey
@@ -93,8 +102,13 @@ impl IdentityTopUpStateTransitionStateValidationV0 for IdentityTopUpTransition {
                 ))
             })?;
 
+        execution_context.add_operation(ExecutionOperation::DoubleSha256);
+        execution_context.add_operation(ExecutionOperation::SignatureVerification(
+            SignatureVerificationOperation::new(KeyType::ECDSA_HASH160),
+        ));
+
         if let Err(e) = signer::verify_hash_signature(
-            &double_sha(singable_bytes),
+            &double_sha(signable_bytes),
             self.signature().as_slice(),
             public_key_hash,
         ) {