+++ packages/rs-drive-abci/tests/strategy_tests/strategy.rs
@@ -174,6 +174,29 @@ pub struct NetworkStrategy {
     pub failure_testing: Option<FailureStrategy>,
     pub query_testing: Option<QueryStrategy>,
     pub verify_state_transition_results: bool,
+    pub max_tx_bytes_per_block: i64,
+}
+
+impl Default for NetworkStrategy {
+    fn default() -> Self {
+        NetworkStrategy {
+            strategy: Default::default(),
+            total_hpmns: 100,
+            extra_normal_mns: 0,
+            quorum_count: 24,
+            upgrading_info: None,
+            core_height_increase: Frequency {
+                times_per_block_range: Default::default(),
+                chance_per_block: None,
+            },
+            proposer_strategy: Default::default(),
+            rotate_quorums: false,
+            failure_testing: None,
+            query_testing: None,
+            verify_state_transition_results: false,
+            max_tx_bytes_per_block: 44800,
+        }
+    }
 }
 
 #[derive(Clone, Debug)]
@@ -845,18 +868,9 @@ impl NetworkStrategy {
                         let owner = current_identities.get(indices[0]).unwrap();
                         let recipient = current_identities.get(indices[1]).unwrap();
 
-                        let fetched_owner_balance = platform
-                            .drive
-                            .fetch_identity_balance(owner.id().to_buffer(), None, platform_version)
-                            .expect("expected to be able to get identity")
-                            .expect("expected to get an identity");
-
                         let state_transition =
                             strategy_tests::transitions::create_identity_credit_transfer_transition(
-                                owner,
-                                recipient,
-                                signer,
-                                fetched_owner_balance - 100,
+                                owner, recipient, signer, 1000,
                             );
                         operations.push(state_transition);
                     }