+++ packages/rs-drive-abci/tests/strategy_tests/verify_state_transitions.rs
@@ -26,6 +26,8 @@ use dpp::state_transition::documents_batch_transition::accessors::DocumentsBatch
 use drive::state_transition_action::document::documents_batch::document_transition::document_base_transition_action::DocumentBaseTransitionActionAccessorsV0;
 use drive::state_transition_action::document::documents_batch::document_transition::document_create_transition_action::DocumentFromCreateTransitionAction;
 use drive::state_transition_action::document::documents_batch::document_transition::document_replace_transition_action::DocumentFromReplaceTransitionAction;
+use drive_abci::execution::types::state_transition_execution_context::StateTransitionExecutionContext;
+use platform_version::DefaultForPlatformVersion;
 
 pub(crate) fn verify_state_transitions_were_or_were_not_executed(
     abci_app: &AbciApplication<MockCoreRPCLike>,
@@ -55,17 +57,25 @@ pub(crate) fn verify_state_transitions_were_or_were_not_executed(
                 // dbg!(batch.transitions().len(), hex::encode(first.base().id()), state.height(), first.to_string());
             }
 
-            let consensus_validation_result =
-                match state_transition.transform_into_action(&platform, false, None) {
-                    Ok(consensus_validation_result) => consensus_validation_result,
-                    Err(e) => {
-                        if expected_validation_errors.contains(&result.code) {
-                            return (state_transition.clone(), None, false);
-                        } else {
-                            panic!("{}", e)
-                        }
+            let mut execution_context =
+                StateTransitionExecutionContext::default_for_platform_version(platform_version)
+                    .expect("expected to get an execution context");
+
+            let consensus_validation_result = match state_transition.transform_into_action(
+                &platform,
+                false,
+                &mut execution_context,
+                None,
+            ) {
+                Ok(consensus_validation_result) => consensus_validation_result,
+                Err(e) => {
+                    if expected_validation_errors.contains(&result.code) {
+                        return (state_transition.clone(), None, false);
+                    } else {
+                        panic!("{}", e)
                     }
-                };
+                }
+            };
 
             if !consensus_validation_result.is_valid() {
                 panic!(
@@ -408,22 +418,12 @@ pub(crate) fn verify_state_transitions_were_or_were_not_executed(
                         platform.state.last_committed_block_info()
                     );
                     if *was_executed {
-                        assert_eq!(
-                            identity
-                                .expect("expected an identity")
-                                .into_partial_identity_info_no_balance(),
-                            PartialIdentity {
-                                id: identity_create_transition.identity_id(),
-                                loaded_public_keys: identity_create_transition
-                                    .public_keys()
-                                    .iter()
-                                    .map(|key| (key.id(), key.clone()))
-                                    .collect(),
-                                balance: None,
-                                revision: Some(0),
-                                not_found_public_keys: Default::default(),
-                            }
-                        )
+                        // other state transitions might have happened in the same block the identity
+                        // was created
+                        let proved_identity = identity
+                            .expect("expected an identity")
+                            .into_partial_identity_info_no_balance();
+                        assert_eq!(proved_identity.id, identity_create_transition.identity_id());
                     } else {
                         //there is the possibility that the state transition was not executed because it already existed,
                         // we can discount that for now in tests
@@ -577,6 +577,8 @@ pub(crate) fn verify_state_transitions_were_or_were_not_executed(
                             &identity_update_transition.identity_id().into_buffer(),
                             None,
                         ),
+                        false,
+                        false,
                         false,
                         platform_version,
                     )